requirements.txt
fastapi
uvicorn
openai
langchain
# faiss-cpu
pymongo
tiktoken
python-dotenv
streamlit
langchain-community
pydantic
sentence-transformers
transformers 
accelerate
langchain_huggingface
qdrant_client